add synthset support in datasets.json
